Updated server for client version 3.0.3.0 r330995 from 15.03.2017 based on branch 'develop+QNDS' from 16.09.2022; - updated opcodes and constants; - the server login works; - you can enter the lobby; - you can create characters; - you can enter world; - not all package structures have been checked, further work on them is required;

+++ AAEmu.Commons/Models/LoginCharacterInfo.cs
@@ -5,7 +5,7 @@ namespace AAEmu.Commons.Models
     public class LoginCharacterInfo : PacketMarshaler
     {
         public uint Id { get; set; }
-        public uint AccountId { get; set; }
+        public ulong AccountId { get; set; }
         public byte GsId { get; set; }
         public string Name { get; set; }
         public byte Race { get; set; }
