+++ AAEmu.Game/Models/Game/Items/Templates/EquipItemTemplate.cs
@@ -11,8 +11,8 @@ namespace AAEmu.Game.Models.Game.Items.Templates
         public int DurabilityMultiplier { get; set; }
         public uint RechargeBuffId { get; set; }
         public int ChargeLifetime { get; set; }
-        public int ChargeCount { get; set; } // does not seem to be actually used anywhere in the DB
+        public short ChargeCount { get; set; } // does not seem to be actually used anywhere in the DB
         public ItemLookConvert ItemLookConvert { get; set; }
-        public uint EquipItemSetId { get; set; }
+        public uint EquipItemSetId { get; set; } // EquipSetId
     }
 }
